Implemented basic classes
After setting up the server/client workflow, have now put in place a simple version of the network event listeners that will be needed to play the game of rummy. This is paired with documentation made on google docs that describes each network event, how they are triggered, and what they do within the game. There is also now a Card object that has a number, suit and an image. The number and suit are assigned when it is created and the image will be deduced from its number and suit.