Server-side stock management
Each card has an id, and are created and stored into an array on both clients, then on the server an arraycontaining each card's id is created. When a client requests the top of the stock pile, the top of the id array is returned, and then the card with the corresponding id is given to the client on their side, this avoids passing card objects over the network.