Drawing and Server/Client changes
Now shows a back facing card where the stock pile sits, the top most card in the discard pile where the discard pile would sit. Now allow both clients to discard a card from their hand and add it to the discard pile, this is managed serverside in terms of ids, but the card objects sit on the local client side. Now shows all cards in the client's hand on the screen face up, and shows the other clients cards but as a back facing card. Finally, have split up the networks send events and network receive events for readability and made it so that functions that simply remove the top card in an array do not require a card argument to function.